A couple of int --> size_t conversions.
git-svn-id: http://svn.atnf.csiro.au/duchamp/trunk@995 450e7745-72da-457f-8a74-b07dc0cda934

+++ src/Cubes/baseline.cc
@@ -47,11 +47,11 @@ namespace duchamp
 
     float *spec     = new float[this->axisDim[2]];
     float *thisBaseline = new float[this->axisDim[2]];
-    int numSpec = this->axisDim[0]*this->axisDim[1];
+    size_t numSpec = this->axisDim[0]*this->axisDim[1];
 
     ProgressBar bar;
     if(this->par.isVerbose()) bar.init(numSpec);
-    for(int pix=0; pix<numSpec; pix++){ // for each spatial pixel...
+    for(size_t pix=0; pix<numSpec; pix++){ // for each spatial pixel...
 
       if(this->par.isVerbose() ) bar.update(pix+1);
